REF: epd/buildsystem-specific now entierly in EPDPlatform.
