+++ okonomiyaki/platforms/tests/test_epd_platform.py
@@ -8,6 +8,7 @@ from ..epd_platform import (
     _guess_epd_platform, EPD_PLATFORM_SHORT_NAMES, applies
 )
 from ..legacy import _SUBDIR
+from ..platform import DARWIN, LINUX, MAC_OS_X, RHEL, WINDOWS
 from .._arch import Arch, X86, X86_64
 
 from .common import (
@@ -237,3 +238,71 @@ class TestGuessEPDPlatform(unittest.TestCase):
         for args, r_platform in examples:
             platform = EPDPlatform._from_spec_depend_data(*args)
             self.assertEqual(platform, r_platform)
+
+    def test_from_epd_platform_string(self):
+        # Given
+        epd_platform_string = "rh5-32"
+
+        # When
+        epd_platform = EPDPlatform.from_epd_string(epd_platform_string)
+
+        # Then
+        platform = epd_platform.platform
+        self.assertEqual(platform.os, LINUX)
+        self.assertEqual(platform.family, RHEL)
+        self.assertEqual(platform.name, RHEL)
+        self.assertEqual(platform.arch, Arch.from_name(X86))
+        self.assertEqual(platform.machine, Arch.from_name(X86))
+
+        # Given
+        epd_platform_string = "win-32"
+
+        # When
+        epd_platform = EPDPlatform.from_epd_string(epd_platform_string)
+
+        # Then
+        platform = epd_platform.platform
+        self.assertEqual(platform.os, WINDOWS)
+        self.assertEqual(platform.family, WINDOWS)
+        self.assertEqual(platform.name, WINDOWS)
+        self.assertEqual(platform.arch, Arch.from_name(X86))
+        self.assertEqual(platform.machine, Arch.from_name(X86))
+
+        # Given
+        epd_platform_string = "osx-64"
+
+        # When
+        epd_platform = EPDPlatform.from_epd_string(epd_platform_string)
+        platform = epd_platform.platform
+
+        # Then
+        self.assertEqual(platform.os, DARWIN)
+        self.assertEqual(platform.family, MAC_OS_X)
+        self.assertEqual(platform.name, MAC_OS_X)
+        self.assertEqual(platform.arch, Arch.from_name(X86_64))
+        self.assertEqual(platform.machine, Arch.from_name(X86_64))
+
+    def test_from_epd_platform_string_invalid(self):
+        # Given
+        # Invalid bitwidth
+        epd_platform_string = "linux-32-1"
+
+        # When/Then
+        with self.assertRaises(OkonomiyakiError):
+            EPDPlatform.from_epd_string(epd_platform_string)
+
+        # Given
+        # Invalid bitwidth
+        epd_platform_string = "osx-63"
+
+        # When/Then
+        with self.assertRaises(OkonomiyakiError):
+            EPDPlatform.from_epd_string(epd_platform_string)
+
+        # Given
+        # Invalid platform basename
+        epd_platform_string = "netbsd-32"
+
+        # When/Then
+        with self.assertRaises(OkonomiyakiError):
+            EPDPlatform.from_epd_string(epd_platform_string)
